refactor: 금액 자료형 BigDecimal로 관리

+++ src/main/java/roomescape/dto/payment/PaymentRequest.java
@@ -1,4 +1,6 @@
 package roomescape.dto.payment;
 
-public record PaymentRequest(String orderId, int amount, String paymentKey) {
+import java.math.BigDecimal;
+
+public record PaymentRequest(String orderId, BigDecimal amount, String paymentKey) {
 }
