+++ src/main/java/roomescape/dto/request/reservation/ReservationRequest.java
@@ -1,6 +1,7 @@
 package roomescape.dto.request.reservation;
 
 import jakarta.validation.constraints.NotNull;
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 public record ReservationRequest(
@@ -9,6 +10,6 @@ public record ReservationRequest(
         long themeId,
         String paymentKey,
         String orderId,
-        int amount
+        BigDecimal amount
 ) {
 }
