+++ src/test/java/roomescape/client/PaymentClientTest.java
@@ -7,6 +7,7 @@ import static org.springframework.test.web.client.match.MockRestRequestMatchers.
 import static org.springframework.test.web.client.match.MockRestRequestMatchers.requestTo;
 import static org.springframework.test.web.client.response.MockRestResponseCreators.withStatus;
 
+import java.math.BigDecimal;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -33,7 +34,7 @@ class PaymentClientTest {
     @DisplayName("결제가 이상없이 승인되면 올바른 응답을 반환한다.")
     @Test
     void paymentSuccess() {
-        PaymentRequest request = new PaymentRequest("randomOrderId", 10000, "randomPaymentKey");
+        PaymentRequest request = new PaymentRequest("randomOrderId", BigDecimal.valueOf(10000), "randomPaymentKey");
         String expectedResponse = """
                 {
                     "paymentKey" : "randomPaymentKey",
@@ -49,14 +50,14 @@ class PaymentClientTest {
         PaymentResponse result = paymentClient.requestPayment(request);
         assertAll(
                 () -> assertThat(result.paymentKey()).isEqualTo("randomPaymentKey"),
-                () -> assertThat(result.totalAmount()).isEqualTo(10000)
+                () -> assertThat(result.totalAmount()).isEqualTo(BigDecimal.valueOf(10000))
         );
     }
 
     @DisplayName("결제 승인 요청이 실패하면 예외를 던진다.")
     @Test
     void paymentFailure() {
-        PaymentRequest request = new PaymentRequest("randomOrderId", 10000, "randomPaymentKey");
+        PaymentRequest request = new PaymentRequest("randomOrderId", BigDecimal.valueOf(10000), "randomPaymentKey");
         String expectedResponse = """
                 {
                     "code": "NOT_FOUND_PAYMENT",
